Rename secs to dsecs; remove old SecsInputs
