+++ docs/api/dsecs.rst
@@ -1,17 +1,17 @@
 swarmpal.toolboxes.secs
 =======================
 
-swarmpal.toolboxes.secs.processes
-*********************************
+swarmpal.toolboxes.dsecs.processes
+**********************************
 
-.. automodule:: swarmpal.toolboxes.secs.processes
+.. automodule:: swarmpal.toolboxes.dsecs.processes
     :members:
     :undoc-members:
     :inherited-members:
     :show-inheritance:
 
-swarmpal.toolboxes.secs.dsecs_algorithms
-*********************************
+swarmpal.toolboxes.dsecs.dsecs_algorithms
+*****************************************
 
 .. automodule:: swarmpal.toolboxes.secs.dsecs_algorithms
     :members:
@@ -19,8 +19,8 @@ swarmpal.toolboxes.secs.dsecs_algorithms
     :inherited-members:
     :show-inheritance:
 
-swarmpal.toolboxes.secs.aux_tools
-*********************************
+swarmpal.toolboxes.dsecs.aux_tools
+**********************************
 
 .. automodule:: swarmpal.toolboxes.secs.aux_tools
     :members:
