+++ src/swarmpal/toolboxes/dsecs/__init__.py
@@ -1,0 +1,7 @@
+from swarmpal.toolboxes.dsecs import aux_tools, dsecs_algorithms, processes
+
+__all__ = (
+    "aux_tools",
+    "dsecs_algorithms",
+    "processes",
+)
