--- src/swarmpal/toolboxes/secs/__init__.py
@@ -1,6 +1,0 @@
-from swarmpal.toolboxes.secs.secs_inputs import SecsInputs, SecsInputSingleSat
-
-__all__ = (
-    "SecsInputSingleSat",
-    "SecsInputs",
-)
